Conflicts:
	Dev/Warewolf.Studio.Views/ServiceTestView.xaml

+++ Dev/Warewolf.Studio.ViewModels/ServiceTestViewModel.cs
@@ -273,6 +273,10 @@ namespace Warewolf.Studio.ViewModels
         {
             if (SelectedServiceTest.IsDirty)
             {
+                if (ShowPopupWhenDuplicates())
+                {
+                    return;
+                }
                 Save(new List<IServiceTestModel> { SelectedServiceTest });
             }
             ServiceTestCommandHandler.RunSelectedTest(SelectedServiceTest, ResourceModel, AsyncWorker);
@@ -490,7 +494,7 @@ namespace Warewolf.Studio.ViewModels
         {
             try
             {
-                if (ShowPoputWhenDuplicates())
+                if (ShowPopupWhenDuplicates())
                 {
                     return;
                 }
@@ -575,7 +579,7 @@ namespace Warewolf.Studio.ViewModels
 
         }
 
-        private bool ShowPoputWhenDuplicates()
+        private bool ShowPopupWhenDuplicates()
         {
             if (HasDuplicates())
             {
@@ -756,10 +760,12 @@ namespace Warewolf.Studio.ViewModels
             {
                 try
                 {
-                    ResourceModel.Environment.ResourceRepository.DeleteResourceTest(ResourceModel.ID, test.TestName);
-                    var testToRemove = _tests.SingleOrDefault(model => model.ParentId == test.ParentId && model.TestName == SelectedServiceTest.TestName);
-                    _tests.Remove(testToRemove); //test
-                    OnPropertyChanged(() => Tests); //test
+                    if (!test.IsNewTest)
+                    {
+                        ResourceModel.Environment.ResourceRepository.DeleteResourceTest(ResourceModel.ID, test.TestName);
+                    }
+                    _tests.Remove(test);
+                    OnPropertyChanged(() => Tests);
                     SelectedServiceTest = null;
                 }
                 catch (Exception ex)
